moved Admin page into its own java file
